+++ workspaces/scorecard/plugins/scorecard-backend-module-github/src/metricProviders/GithubOpenPRsProvider.ts
@@ -1,0 +1,41 @@
+/*
+ * Copyright Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Metric } from '@red-hat-developer-hub/backstage-plugin-scorecard-common';
+import { MetricProvider } from '@red-hat-developer-hub/backstage-plugin-scorecard-node';
+
+export class GithubOpenPRsProvider implements MetricProvider<'number'> {
+  getProviderDatasourceId(): string {
+    return 'github';
+  }
+
+  getProviderId() {
+    return 'github.open-prs';
+  }
+
+  getMetric(): Metric<'number'> {
+    return {
+      id: this.getProviderId(),
+      title: 'Github open PRs',
+      type: 'number',
+      history: true,
+    };
+  }
+
+  async calculateMetric(): Promise<number> {
+    return 42;
+  }
+}