+++ workspaces/ai-integrations/packages/backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # backend
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [551bffa]
+  - @red-hat-developer-hub/backstage-plugin-catalog-techdoc-url-reader-backend@0.2.2
+
 ## 0.0.6
 
 ### Patch Changes